# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DER-EMS API",
        "contact": {},
        "license": {}
    },
    "basePath": "/api",
    "paths": {
...
        "/{gwid}/devices/charge-info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get demand charge information by token, gateway UUID and startTime",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demand charge"
                ],
                "summary": "Show the demand charge information",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Input user's access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gateway UUID",
                        "name": "gwid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "startTime",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.ChargeInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/{gwid}/devices/demand-state": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get demand details by token, gateway UUID, startTime and endTime",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demand charge"
                ],
                "summary": "Show today's demand details",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Input user's access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Gateway UUID",
                        "name": "gwid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "endTime",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "example": "UTC time in ISO-8601",
                        "name": "startTime",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/services.DemandStateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "services.ChargeInfoResponse": {
            "type": "object",
            "properties": {
                "gridContractPowerAC": {
                    "type": "number"
                },
                "gridIsPeakShaving": {
                    "type": "integer"
                },
                "gridPowerCost": {
                    "type": "number"
                },
                "gridPowerCostLastMonth": {
                    "type": "number"
                },
                "gridPowerCostSavings": {
                    "type": "number"
                },
                "gridProducedAveragePowerAC": {
                    "type": "number"
                }
            }
        },
        "services.DemandStateResponse": {
            "type": "object",
            "properties": {
                "gridContractPowerAC": {
                    "type": "number"
                },
                "gridLifetimeEnergyACDiffToPowers": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "timestamps": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}